refactor(net-stubbing): move cy.intercept core from network-interception
